+++ internal/es/estest/common.go
@@ -1,3 +1,4 @@
+//nolint:govet // this is a test file
 package estest
 
 import (